Fix aspect-drivers precompute failures and mention-count overflow
- scanCategoryPlatform was filtering on sentiment_score <> 0, a column
  that does not exist on youtube_comments, reddit_posts, or
  instagram_posts; this caused the entire nightly precompute job to fail
  silently, leaving the aspect_drivers_agg table empty. Fixed by
  filtering on sentiment_category IS NOT NULL instead.
- scanX SQL tightened to BETWEEN 1 AND 100 (enforces the valid range at
  the DB level); the Java guard updated to match (raw < 1.0 || raw > 100.0)
  so out-of-range values from direct DB writes are also rejected.
- rs.getInt("n") replaced with rs.getLong("n") when reading
  SUM(mention_count) — Postgres returns BIGINT for aggregate sums,
  which rs.getInt silently truncates above ~2.1 billion. The downstream
  cast in buildStrengthsWeaknesses is updated to long accordingly.

+++ src/main/java/com/lit/fire/flame/ViralSeedController.java
@@ -299,7 +299,7 @@ public class ViralSeedController {
                             .computeIfAbsent(rs.getString("platform"), k -> new HashMap<>())
                             .computeIfAbsent(rs.getString("aspect"), k -> new double[2]);
                     acc[0] += rs.getDouble("s");
-                    acc[1] += rs.getInt("n");
+                    acc[1] += rs.getLong("n");
                 }, args);
 
         // Overall view: merge every platform's aspect aggregates by summing sums and counts.
@@ -355,7 +355,7 @@ public class ViralSeedController {
 
         for (Map.Entry<String, double[]> e : aspectMap.entrySet()) {
             double sum = e.getValue()[0];
-            int n = (int) e.getValue()[1];
+            long n = (long) e.getValue()[1];
             if (n < MIN_ASPECT_MENTIONS) continue;
 
             double avg = sum / n;
